requirements.txt
streamlit
tensorflow==2.15.0
numpy
Pillow
reportlab
qrcode
huggingface_hub
requests
